Home work#53

+++ src/main/java/kg/job/jobsearch/controller/ResumeController.java
@@ -1,12 +1,12 @@
 package kg.job.jobsearch.controller;
 
-import kg.job.jobsearch.dto.ResumeCreateDto;
-import kg.job.jobsearch.dto.ResumeUpdateDto;
+import kg.job.jobsearch.dto.create.ResumeCreateDto;
+import kg.job.jobsearch.dto.update.ResumeUpdateDto;
 import kg.job.jobsearch.dto.ResumesDto;
-import kg.job.jobsearch.exception.ContactsInfoNotFoundException;
-import kg.job.jobsearch.exception.EducationNotFoundException;
-import kg.job.jobsearch.exception.ResumeNotFoundException;
-import kg.job.jobsearch.exception.WorkExperienceInfoNotFoundException;
+import kg.job.jobsearch.exception.notFoundException.ContactsInfoNotFoundException;
+import kg.job.jobsearch.exception.notFoundException.EducationNotFoundException;
+import kg.job.jobsearch.exception.notFoundException.ResumeNotFoundException;
+import kg.job.jobsearch.exception.notFoundException.WorkExperienceInfoNotFoundException;
 import kg.job.jobsearch.service.ResumeService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
@@ -25,7 +25,7 @@ public class ResumeController {
     }
 
     @GetMapping("/search/by-category")
-    public List<ResumesDto> getResumesByCategory(@RequestParam Integer category) throws ResumeNotFoundException {
+    public List<ResumesDto> getResumesByCategory(@RequestParam Long category) throws ResumeNotFoundException {
         return resumeService.getResumeByCategory(category);
     }
